feat: Add SR toolGroup creation on modeEnter
