+++ extensions/cornerstone-3d/src/commandsModule.js
@@ -112,10 +112,11 @@ const commandsModule = ({ servicesManager }) => {
           continue;
         }
 
-        // only use the toolGroup viewport
-        if (!toolGroupViewportIds.includes(viewport.viewportId)) {
-          continue;
-        }
+        // Commenting this out, since we should be able to set the tool to be
+        // active on a toolGroup whose viewports have not been added yet.
+        // if (!toolGroupViewportIds.includes(viewport.viewportId)) {
+        //   continue;
+        // }
 
         // Find the current active tool and set it to be passive
         const activeTool = toolGroup.getActivePrimaryMouseButtonTool();
