+++ extensions/cornerstone-dicom-sr/src/tools/toolNames.ts
@@ -1,0 +1,9 @@
+import SRLengthTool from './annotationTools/SRLengthTool';
+import DICOMSRDisplayTool from './DICOMSRDisplayTool';
+
+const toolNames = {
+  SRLengthTool: SRLengthTool.toolName,
+  DICOMSRDisplayTool: DICOMSRDisplayTool.toolName,
+};
+
+export default toolNames;
